Make DOM queries fully lazy
Previously DOM queries weren't fully lazy -- when doing an indexing operation, we would execute a DOM query to get it results, index into them, and create a new root for the child DOM query.

This change defers that until actually executing the child query. So, rather than a DOMQuery's data consisting of a root element and a single selector, it consists of a root element and a selector array that trackes the selectors/indices involves in the query so it can defer any evaluation (querying the actual DOM) until the query is actually executed.

This allows more flexibility in creating PageObjects before the DOM is populated, and also paves the way for async DOM queries (to support webdriver-like test environments).