Fix type errors in Ember tests
#89 introduced some type errors in the Ember tests that I didn't notice because noEmitOnError was set to false. So set it to true and fix up the type errors.

In fixing the type errors, I realized that wrapping `selector()` has gotten a good bit more complex because `selector()`'s type is more complex, so I re-implemented the test to define a generic `wrapSelector()` function that can be used to produce a wrapper that transforms the arguments before passing them to `selector()`. I did the same for `globalSelector()` (whose type is even more complex). We could potentially put these in the library and export them to make them available to consumers, but I want to give them a little more time and get some feedback from the folks (or maybe just one person?) that wraps `selector()` before promoting them to a "public API."