Better shadow DOM implementation
The previous shadow DOM support implementation broke the model a bit. It basically made anything that could be an Element also allowed to be a DocumentFragment. But this wasn't really the intent -- this library is meant to find Elements, not DocumentFragments, and really we just need DocumentFragments to be able to be the _parents_ of PageObjects, not their target.

So adjust the code and types in DOMQuery to capture the fact that DocumentFragments can be query roots, but not query results, and then update all the other types to reflect this as well.